+++ src/main/java/com/keunsori/keunsoriserver/domain/reservation/dto/requset/ReservationCreateRequest.java
@@ -1,4 +1,4 @@
-package com.keunsori.keunsoriserver.domain.reservation.dto;
+package com.keunsori.keunsoriserver.domain.reservation.dto.requset;
 
 import com.keunsori.keunsoriserver.domain.member.domain.Member;
 import com.keunsori.keunsoriserver.domain.reservation.domain.Reservation;
@@ -6,8 +6,8 @@ import com.keunsori.keunsoriserver.domain.reservation.domain.vo.ReservationType;
 import com.keunsori.keunsoriserver.domain.reservation.domain.vo.Session;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import java.sql.Time;
 import java.time.LocalDate;
+import java.time.LocalTime;
 
 public record ReservationCreateRequest(
         ReservationType reservationType,
@@ -15,9 +15,9 @@ public record ReservationCreateRequest(
         @Schema(example = "2025-01-01", type = "string")
         LocalDate reservationDate,
         @Schema(example = "21:00:00", type = "string")
-        Time reservationStartTime,
+        LocalTime reservationStartTime,
         @Schema(example = "21:00:00", type = "string")
-        Time reservationEndTime
+        LocalTime reservationEndTime
 ) {
     public Reservation toEntity(Member member) {
         return Reservation.builder()
