+++ src/main/java/com/keunsori/keunsoriserver/domain/reservation/dto/response/ReservationResponse.java
@@ -1,18 +1,17 @@
-package com.keunsori.keunsoriserver.domain.reservation.dto;
+package com.keunsori.keunsoriserver.domain.reservation.dto.response;
 
 import com.keunsori.keunsoriserver.domain.reservation.domain.Reservation;
 import com.keunsori.keunsoriserver.domain.reservation.domain.vo.ReservationType;
-import com.keunsori.keunsoriserver.domain.reservation.domain.vo.Session;
 
-import java.sql.Time;
 import java.time.LocalDate;
+import java.time.LocalTime;
 
 public record ReservationResponse(
         Long reservationId,
-        Session session,
+        String session,
         LocalDate reservationDate,
-        Time reservationStartTime,
-        Time reservationEndTime,
+        LocalTime reservationStartTime,
+        LocalTime reservationEndTime,
         ReservationType reservationType,
         Long reservationMemberId,
         String reservationMemberName
@@ -21,7 +20,7 @@ public record ReservationResponse(
     public static ReservationResponse of(Reservation reservation) {
         return new ReservationResponse(
                 reservation.getId(),
-                reservation.getSession(),
+                reservation.getSession().name().toLowerCase(),
                 reservation.getDate(),
                 reservation.getStartTime(),
                 reservation.getEndTime(),
